Fix to handle hex numeric XML entities
If the parser encounters a hex entity, eg &#xA;, entitySwap correctly
identifies it as a numeric entity but causes a 'bad argument' error as
Lua requires hex numbers to be prefixed with '0x', i.e. '0XA' not 'xA'.

By prepending '0' to the numeric portion of the entity, hex values are
correctly formed and decimal values are still valid, e.g. '0xA' & '010'